Selector: Rename rcombinators to rleadingCombinator
