+++ src/selector.js
@@ -10,7 +10,7 @@ import rtrim from "./var/rtrim.js";
 import isIE from "./var/isIE.js";
 import identifier from "./selector/var/identifier.js";
 import booleans from "./selector/var/booleans.js";
-import rcombinators from "./selector/var/rcombinators.js";
+import rleadingCombinator from "./selector/var/rleadingCombinator.js";
 import rdescend from "./selector/var/rdescend.js";
 import rsibling from "./selector/var/rsibling.js";
 import matches from "./selector/var/matches.js";
@@ -159,7 +159,7 @@ function find( selector, context, results, seed ) {
 				// as such selectors are not recognized by querySelectorAll.
 				// Thanks to Andrew Dupont for this technique.
 				if ( nodeType === 1 &&
-					( rdescend.test( selector ) || rcombinators.test( selector ) ) ) {
+					( rdescend.test( selector ) || rleadingCombinator.test( selector ) ) ) {
 
 					// Expand context for sibling selectors
 					newContext = rsibling.test( selector ) &&
